od-ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
scikit-learn==1.6.0
transformers==4.47.0
torch==2.6.0
huggingface-hub==0.27.0
pdfplumber==0.11.0
python-docx==1.1.2
httpx==0.27.0
pydantic==2.10.6
python-multipart==0.0.9
numpy==2.2.0
